add patch endpoint for update isComplete

+++ backend/routers/todos.js
@@ -83,7 +83,28 @@ router.put('/:id', async (req, res) => {
   }
 });
 
+router.patch('/:id', async (req, res) => {
+  const todo = await Todo.findById(req.params.id);
+
+  if (!todo) return res.status(404).send('Todo not found...');
+
+  try {
+    const updateTodo = await Todo.findByIdAndUpdate(req.params.id, {
+      isComplete: !todo.isComplete,
+    });
+
+    res.send(updateTodo);
+  } catch (error) {
+    res.status(500).send(error.message);
+    console.log(error.message);
+  }
+});
+
 router.delete('/:id', async (req, res) => {
+  const todo = await Todo.findById(req.params.id);
+
+  if (!todo) return res.status(404).send('Todo not found...');
+
   try {
     const deleteTodo = await Todo.findByIdAndDelete(req.params.id);
